Move to arg model
